Remove deprecated code
This change removes usage of deprecated code in the codebase, either
replacing it with the recommended replacement, or the standard
replacement.

So, MarkdownConverterInterface was replaced with the recommended
replacement of ConverterInterface, and the applicable function call was
replaced. Then, getMockForAbstractClass, which is deprecated without
replacement in PHPUnit 12 was replaced with createMock.