Update actions/cache to v4
On attempting to run the workflow, I encountered the following error:
"Missing download info for actions/cache@v2". After some searching on
the error, the recommended fix was to update from version 2 to 4. So,
this change is being made to test the solution and, all being well,
correct the issue.